# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Projet API",
    "description": "Proof-of-work hiring infrastructure.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/roles/clusters": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "List Clusters",
        "operationId": "list_clusters_roles_clusters_get",
        "parameters": [
          {
            "name": "projet_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClusterOut"
                  },
                  "title": "Response List Clusters Roles Clusters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roles/search": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "Search Roles",
        "description": "FR-043 \u2014 search runs across names and aliases, so \"data analyst\" or \"BI\"\nlands on Data Analytics rather than creating a near-duplicate.",
        "operationId": "search_roles_roles_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Q"
            }
          },
          {
            "name": "projet_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleSummary"
                  },
                  "title": "Response Search Roles Roles Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roles/{role_id}/implications": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "Role Implications",
        "description": "FR-053 \u2014 what picking this role commits them to, before they commit.\n\nThe deliverable, the sources, the two role-specific criteria and the\nlow-friction asks. Seeing the artifact spec now is what prevents discovering\nafter the cohort ran that it was not what they wanted.",
        "operationId": "role_implications_roles__role_id__implications_get",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Role Id"
            }
          },
          {
            "name": "projet_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleImplications"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClusterOut": {
        "properties": {
          "cluster": {
            "type": "string",
            "title": "Cluster"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleSummary"
            },
            "type": "array",
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "cluster",
          "roles"
        ],
        "title": "ClusterOut",
        "description": "FR-052 \u2014 cluster first, then role.\n\nA flat list of 75 is unusable; picking \"Finance\" then \"Venture Capital\" is\ntwo taps and prevents mis-tagging."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleImplications": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/RoleSummary"
          },
          "default_deliverable": {
            "type": "string",
            "title": "Default Deliverable"
          },
          "universal_memo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Universal Memo"
          },
          "public_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Public Sources"
          },
          "student_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Student Tools"
          },
          "judging_criteria": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Judging Criteria"
          },
          "company_asks_easy": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Company Asks Easy"
          },
          "delivery_risk_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Risk Note"
          }
        },
        "type": "object",
        "required": [
          "role",
          "default_deliverable",
          "public_sources",
          "student_tools",
          "judging_criteria",
          "company_asks_easy"
        ],
        "title": "RoleImplications",
        "description": "FR-053 \u2014 what the company sees the moment they pick a role.\n\nThis is where the programme stops being abstract for the buyer, and it\nprevents the common failure where a company picks a role whose deliverable\nis not what they actually wanted."
      },
      "RoleSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "cluster": {
            "type": "string",
            "title": "Cluster"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "cluster"
        ],
        "title": "RoleSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}